Move collection filter to new module
